Refactor AI client interface and add metrics tracking with statistical analysis
- Refactor AiClient to use sendMessageWithMetrics as primary method
- Move JSON response logic to separate JsonResponseInteractor
- Add MetricsCollector for tracking response metrics (duration, tokens)
- Display real-time performance comparison with average and median (50p)
- Update max message length from 1000 to 5000 characters
- Add /json command for structured JSON responses
- Extract actual token counts from all AI providers (Claude, HuggingFace, Yandex)
- Add multi-message conversation support with ConversationMessage
- Remove unnecessary runBlocking calls in suspend contexts

+++ src/main/kotlin/com/com/bot/DifferentModelsInteractor.kt
@@ -75,7 +75,7 @@ class DifferentModelsInteractor(
         )
     }
 
-    private fun formatMetrics(durationMs: Long, promptTokens: Int, completionTokens: Int, totalTokens: Int): String {
+    private fun formatMetrics(durationMs: Long, promptTokens: Long, completionTokens: Long, totalTokens: Long): String {
         val durationSeconds = durationMs / 1000.0
         return """
             📊 *Metrics:*
